change: Allowing for user-specified timeframes
In addition to this:

Added functionality to handle the use case when only part of a frame is
in the timespan the user is filtering by. In this implementation, the
part of the frame that is in the timespan will be included in the
results.

For example, if a frame starts at 23:30 and ends on 00:30 the next day,
the report for the previous day will include the first 30 minutes of
that frame, and the report for the current day will include the last 30
minutes of the frame.